% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-conc-ellipse.R
\name{fviz_gda_conc_ellipse}
\alias{fviz_gda_conc_ellipse}
\title{Title}
\usage{
fviz_gda_conc_ellipse(
  res_gda,
  level = 0.8647,
  alpha = 0.1,
  colour = "black",
  linetype = "dashed",
  density = FALSE,
  fill = NA,
  axes = 1:2,
  scale_size = 1,
  title = "GDA individuals plot",
  plot_modif_rates = TRUE,
  axis_lab_name = "Achse",
  labels = NULL
)
}
\arguments{
\item{res_gda}{GDA result.}

\item{level}{ellipse level (default 86.47\%).}

\item{alpha}{opacity level (default 0.1).}

\item{colour}{ellipse border colour.}

\item{linetype}{ellipse edge linetype.}

\item{density}{show density contours (boolean).}

\item{fill}{ellipse fill colour.}

\item{axes}{the GDA dimensions to plot.}

\item{scale_size}{scale minimal point size.}

\item{title}{the plot title.}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{axis_lab_name}{name of axis label.}

\item{labels}{label axes (vector of length 4; left, right, top, bottom).}
}
\value{
ggplot2 GDA visualisation with concentration ellipse.
}
\description{
Title
}
