% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-ellipses.R
\name{fviz_gda_trajectory_ellipses}
\alias{fviz_gda_trajectory_ellipses}
\title{Visualization of trajectory structuring factor ellipses.}
\usage{
fviz_gda_trajectory_ellipses(
  res_gda,
  df_var_quali,
  var_quali,
  axes = 1:2,
  impute = TRUE,
  time_point_names = NULL,
  ind_points = TRUE,
  concentration_ellipse = TRUE,
  title = NULL,
  plot_modif_rates = TRUE,
  alpha = 0.15,
  select = NULL,
  select_facet = TRUE,
  labels = NULL,
  xlim = NULL,
  ylim = NULL,
  axes_annotate_alpha = 0.3,
  complete_obs = FALSE,
  facet_title_size = 14
)
}
\arguments{
\item{res_gda}{MCA result (rownames have to be questionnaire IDs including time number, e.g. 87654_1).}

\item{df_var_quali}{data frame containing one qualitative variable (with IDs as rownames).}

\item{var_quali}{name of the structuring variable.}

\item{axes}{the axes to plot.}

\item{impute}{use imputation for missing data.}

\item{time_point_names}{vector containing the name of the time points.}

\item{ind_points}{show individuals (boolean).}

\item{concentration_ellipse}{plot confidence ellipses (boolean).}

\item{title}{title of the plot.}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{alpha}{ellipse fill alpha.}

\item{select}{choose cluster/category.}

\item{select_facet}{facet clusters/categories (boolean.)}

\item{labels}{label axes (vector of length 4; left, right, top, bottom).}

\item{xlim}{x limits.}

\item{ylim}{y limits.}

\item{axes_annotate_alpha}{alpha value of axes annotations.}

\item{complete_obs}{plot only complete observations (boolean).}

\item{facet_title_size}{size of the facet stripe title (numeric).}
}
\value{
ggplot2 visualization.
}
\description{
Visualization of trajectory structuring factor ellipses.
}
