% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-quali.R
\name{fviz_gda_trajectory_quali}
\alias{fviz_gda_trajectory_quali}
\title{Visualize trajectories and structuring factors.}
\usage{
fviz_gda_trajectory_quali(
  res_gda,
  df_var_quali,
  var_quali,
  var_quali_select = NULL,
  axes = 1:2,
  ind_labels = FALSE,
  title = NULL,
  time_point_names = NULL,
  select = list(name = NULL, within_inertia = NULL, case = NULL),
  impute = TRUE,
  plot_modif_rates = TRUE,
  labels = NULL,
  xlim = NULL,
  ylim = NULL,
  axes_annotate_alpha = 0.3
)
}
\arguments{
\item{res_gda}{MCA result (rownames have to be questionnaire IDs including time number, e.g. 87654_1).}

\item{df_var_quali}{data frame containing one qualitative variable (with IDs as rownames).}

\item{var_quali}{name of the structuring variable.}

\item{var_quali_select}{the name of the selected categories/clusters.}

\item{axes}{the axes to plot.}

\item{ind_labels}{plot labels (boolean).}

\item{title}{the plot title.}

\item{time_point_names}{vector containing the name of the time points.}

\item{select}{select vector of names, within_inertia of individuals selection (within_inertia: vector containing the number of high variation and low variationindividuals) or case (vector containing NULL, complete, or incomplete).}

\item{impute}{use imputation for missing data.}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{labels}{label axes (vector of length 4; left, right, top, bottom).}

\item{xlim}{numeric vector of 2.}

\item{ylim}{numeric vector of 2.}

\item{axes_annotate_alpha}{alpha value of axes annotations.}
}
\value{
ggplot2 visualization.
}
\description{
Visualize trajectories and structuring factors.
}
