% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fviz-gda-trajectory-sample.R
\name{fviz_gda_trajectory_sample}
\alias{fviz_gda_trajectory_sample}
\title{Visualization of the separated concentration ellipses of the sample.}
\usage{
fviz_gda_trajectory_sample(res_gda, time_point_names = NULL,
  open_sans = TRUE, axes = 1:2, ind_points = TRUE,
  concentration_ellipse = TRUE, complete = TRUE, title = NULL,
  plot_modif_rates = TRUE, alpha = 0.15, axis_lab_name = "Achse",
  axes_annotate_alpha = 0.3, labels = NULL, legend_x = 0.12,
  legend_y = 0.9)
}
\arguments{
\item{res_gda}{MCA result (rownames have to be questionnaire IDs including time number, e.g. 87654_1).}

\item{time_point_names}{vector containing the name of the time points.}

\item{open_sans}{use Open Sans font (boolean).}

\item{axes}{the axes to plot.}

\item{ind_points}{show individuals (boolean).}

\item{concentration_ellipse}{show or hide overall concentration ellipse (boolean).}

\item{complete}{plot only complete cases (boolean).}

\item{title}{title of the plot}

\item{plot_modif_rates}{plot modified rates instead of eigenvalue percentage (boolean).}

\item{alpha}{ellipse fill alpha.}

\item{axis_lab_name}{name of axis label.}

\item{axes_annotate_alpha}{alpha value of axes annotations.}

\item{labels}{label axes (vector of length 4; left, right, top, bottom).}

\item{legend_x}{x position of legend.}

\item{legend_y}{y position of legend.}
}
\value{
ggplot2 visualization.
}
\description{
Visualization of the separated concentration ellipses of the sample.
}
