% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Visualize a barplot.}
\usage{
plot_barplot(df_origin, df_var, sort = FALSE, bar_abs_size = 5.5,
  bar_rel_size = 4.5, axis_label_size = 12, axis_cat_size = 16,
  show_missing = TRUE, digits = 1, open_sans = TRUE)
}
\arguments{
\item{df_origin}{source data farme (tibble).}

\item{df_var}{categorical variable name.}

\item{sort}{sort bars (boolean).}

\item{bar_abs_size}{size of absolute values in plot.}

\item{bar_rel_size}{size of relative values in plot.}

\item{axis_label_size}{size of y axis labels.}

\item{axis_cat_size}{size of x axis labels.}

\item{show_missing}{include missing values in plot or not (boolean).}

\item{digits}{amount of label value digits.}

\item{open_sans}{use Open Sans font.}
}
\value{
ggplot2 barplot.
}
\description{
Visualize a barplot.
}
