% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology-meaning.R
\name{plot_places_chronology_meaning}
\alias{plot_places_chronology_meaning}
\title{Plot place chronologies map structure and zoom in.}
\usage{
plot_places_chronology_meaning(
  data,
  id,
  weekday = "all",
  size_range = NULL,
  colour_path = "black",
  size_path = 2,
  alpha_path = 0.25,
  alpha_points = 1,
  linetype_path = "solid",
  title = NULL,
  axis_label = FALSE,
  print_place_duration = TRUE,
  open_sans = TRUE,
  exclude_sleep = TRUE,
  facets = FALSE,
  facets_na = FALSE,
  exclude = NULL,
  meanings = NULL,
  map = FALSE,
  map_zoom = 10,
  graph = TRUE,
  area_fill = "white",
  area_colour = "black",
  area_alpha = 0,
  area_size = 2,
  area_linetype = "solid",
  area_label_fontsize = c(12, 10)
)
}
\arguments{
\item{data}{a data frame (columns: ID, day, duration, place, address, lon, lat, prop_duration).}

\item{id}{vector, which contains questionnaire ids. Choosa "all" to compute all ids.}

\item{weekday}{vector, which contains the weekday to plot.}

\item{size_range}{specify the size for visualizatipn of duration.}

\item{colour_path}{sepcify the path line colour.}

\item{size_path}{specify the path line size.}

\item{alpha_path}{specify the path line alpha value [0:1].}

\item{alpha_points}{specify the point alpha value [0:1].}

\item{linetype_path}{specify the linetype of the path line.}

\item{title}{title of the plot.}

\item{axis_label}{show or hide axis labels (boolean).}

\item{print_place_duration}{print place overall duration (hours).}

\item{open_sans}{use Open Sans font (boolean).}

\item{exclude_sleep}{exclude sleep duration (boolean).}

\item{facets}{plot facets (boolean).}

\item{facets_na}{auto remove places without meanings in facets (boolean).}

\item{exclude}{exclude specific places from the plot (vector).}

\item{meanings}{give places a meaning for grouping (vector).}

\item{map}{use map background (boolean).}

\item{map_zoom}{map zoom level.}

\item{graph}{plot graph (boolean).}

\item{area_fill}{fill colour of meaning area.}

\item{area_colour}{line colour of meaning area.}

\item{area_alpha}{alpha of meaning area.}

\item{area_size}{size of meaning area.}

\item{area_linetype}{linetype of meaning area.}

\item{area_label_fontsize}{area label fontsize (vector).}
}
\value{
ggplot2 visualization of place chronology data.
}
\description{
Plot place chronologies map structure and zoom in.
}
