% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology-path.R
\name{plot_places_chronology_path}
\alias{plot_places_chronology_path}
\title{Plot place chronologies activity paths.}
\usage{
plot_places_chronology_path(data, id, recodeded_places)
}
\arguments{
\item{data}{a data frame (columns: ID, day, duration, place, address, lon, lat, prop_duration).}

\item{id}{vector, which contains questionnaire ids. Choose "all" to compute all ids.}

\item{recodeded_places}{recode leves of place labels by named vector.}
}
\value{
ggplot2 visualization of place chronology path.
}
\description{
Plot place chronologies activity paths.
}
