% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-pattern-profile.R
\name{plot_time_pattern_profile}
\alias{plot_time_pattern_profile}
\title{Plot average time pattern profiles.}
\usage{
plot_time_pattern_profile(
  data_tp,
  id = "all",
  ncol = 4,
  fluid = FALSE,
  open_sans = TRUE,
  facet_label_y = 0.95,
  facet_label_sep = 0.05
)
}
\arguments{
\item{data_tp}{data frame containing questionnaire_id, kml3d results and time pattern data.}

\item{id}{time pattern to plot.}

\item{ncol}{facet columns.}

\item{fluid}{should be static bars or fluid lines visualized (boolean).}

\item{open_sans}{use Open Sans font family (boolean).}

\item{facet_label_y}{adjust facet labels y position.}

\item{facet_label_sep}{adjust facet sublabel seperation distance.}
}
\value{
ggplot2 avgerage time pattern profile plot.
}
\description{
Plot average time pattern profiles.
}
