% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Visualize a barplot.}
\usage{
plot_barplot(
  df_origin,
  df_var,
  sort = FALSE,
  bar_abs_size = 3.5,
  bar_rel_size = 3,
  axes_rel_small = 0.6,
  show_missing = TRUE,
  digits = 1,
  flip = FALSE
)
}
\arguments{
\item{df_origin}{source data farme (tibble).}

\item{df_var}{categorical variable name.}

\item{sort}{sort bars (boolean).}

\item{bar_abs_size}{size of absolute values in plot.}

\item{bar_rel_size}{size of relative values in plot.}

\item{axes_rel_small}{relative value for small axes text (labels, titles …).}

\item{show_missing}{include missing values in plot or not (boolean).}

\item{digits}{amount of label value digits.}

\item{flip}{flip axes (boolean).}
}
\value{
ggplot2 barplot.
}
\description{
Visualize a barplot.
}
