% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-places-chronology-meaning.R
\name{plot_places_chronology_meaning}
\alias{plot_places_chronology_meaning}
\title{Plot place chronologies map structure and zoom in.}
\usage{
plot_places_chronology_meaning(
  data,
  id,
  weekday = "all",
  size_range = NULL,
  colour_path = "black",
  size_path = 2,
  alpha_path = 0.25,
  alpha_points = 1,
  linetype_path = "solid",
  title = NULL,
  axis_label = FALSE,
  print_place_duration = TRUE,
  exclude_sleep = TRUE,
  facets = FALSE,
  facets_include_place = NULL,
  facets_include_all = FALSE,
  exclude_na = FALSE,
  exclude = NULL,
  exclude_meaning = NULL,
  meanings = NULL,
  map = FALSE,
  map_zoom = 10,
  map_add_x = 0.2,
  map_add_y = 0.1,
  graph = TRUE,
  area_fill = "white",
  area_colour = "black",
  area_alpha = 0,
  area_size = 1.5,
  area_linetype = "solid",
  area_label_fontsize = c(12, 10),
  area_buffer = 10,
  map_scalebar = TRUE,
  map_scalebar_location = "topright",
  map_scalebar_text_size = 4.5,
  map_scalebar_box_size = 0.015,
  map_scalebar_border_size = 0.85,
  map_scalebar_dist = 1,
  map_scalebar_text_dist = 0.02,
  map_scalebar_unit_pos_dist = 0.5
)
}
\arguments{
\item{data}{a data frame (columns: ID, day, duration, place, address, lon, lat, prop_duration).}

\item{id}{vector, which contains questionnaire ids. Choose "all" to compute all ids.}

\item{weekday}{vector, which contains the weekday to plot.}

\item{size_range}{specify the size for visualizatipn of duration.}

\item{colour_path}{sepcify the path line colour.}

\item{size_path}{specify the path line size.}

\item{alpha_path}{specify the path line alpha value [0:1].}

\item{alpha_points}{specify the point alpha value [0:1].}

\item{linetype_path}{specify the linetype of the path line.}

\item{title}{title of the plot.}

\item{axis_label}{show or hide axis labels (boolean).}

\item{print_place_duration}{print place overall duration (hours).}

\item{exclude_sleep}{exclude sleep duration (boolean).}

\item{facets}{plot facets (boolean).}

\item{facets_include_place}{explicit include places in facets (vector).}

\item{facets_include_all}{include all place names in facet plot (boolean).}

\item{exclude_na}{drop NA places (boolean).}

\item{exclude}{exclude specific places from the plot (vector).}

\item{exclude_meaning}{meanings to be excluded (vector).}

\item{meanings}{give places a meaning for grouping (vector).}

\item{map}{use map background (boolean).}

\item{map_zoom}{map zoom level.}

\item{map_add_x}{adjust map x area.}

\item{map_add_y}{adjust map y area.}

\item{graph}{plot graph (boolean).}

\item{area_fill}{fill colour of meaning area.}

\item{area_colour}{line colour of meaning area.}

\item{area_alpha}{alpha of meaning area.}

\item{area_size}{size of meaning area.}

\item{area_linetype}{linetype of meaning area.}

\item{area_label_fontsize}{area label fontsize (vector).}

\item{area_buffer}{The size of the region around the mark where labels cannot be placed (in mm).}

\item{map_scalebar}{show a scale bar (boolean).}

\item{map_scalebar_location}{location of the scalebar.}

\item{map_scalebar_text_size}{size of the scale text.}

\item{map_scalebar_box_size}{size of the box.}

\item{map_scalebar_border_size}{size of the border.}

\item{map_scalebar_dist}{displayed disctance.}

\item{map_scalebar_text_dist}{distance between box and text.}

\item{map_scalebar_unit_pos_dist}{add space between scalebar values and unit.}
}
\value{
ggplot2 visualization of place chronology data.
}
\description{
Plot place chronologies map structure and zoom in.
}
